ApiKeyAuthenticator instead of guard (#45)
* ApiKeyAuthenticator instead of guard

* Solved phpcs error
